back-server/final_pjt/requirements.txt
asgiref==3.6.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.4
cryptography==40.0.2
decorator==5.0.7
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==3.2
django-allauth==0.52.0
django-bootstrap-v5==1.0.2
django-cors-headers==4.0.0
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.14.0
Faker==8.1.1
idna==3.4
importlib-metadata==2.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.3
oauthlib==3.2.2
parso==0.8.2
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.18
pycparser==2.21
Pygments==2.8.1
PyJWT==2.7.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.3
text-unidecode==1.3
traitlets==5.0.5
urllib3==2.0.2
wcwidth==0.2.5
zipp==3.4.1
